create hypershift openshift-controller-manager
