+++ pkg/cmd/openshift-controller-manager/controller/apps.go
