+++ pkg/cmd/openshift-controller-manager/controller/autoscaling.go
