+++ pkg/cmd/openshift-controller-manager/controller/build.go
