+++ pkg/cmd/openshift-controller-manager/controller/config.go
