+++ pkg/cmd/openshift-controller-manager/controller/image.go
