+++ pkg/cmd/openshift-controller-manager/controller/template.go
