+++ pkg/cmd/openshift-controller-manager/controller/unidling.go
